Converted CLI to use tr-31 protocol (mostly)
Effectively, modifying the identity is no longer a single-step process.
It now has three steps: create transaction, sign transaction N times
(enough to satisfy a policy), and apply the transaction.

This represented some challenges to the interface. In the cases where
the user is managing their own identity, basically not much should
change at all. In the case where a group identity is being managed, we
now have standard `-s/--stage` and `-k/--sign-with` args that allows us
to stage the transaction after signing. This means it lives in a special
area alongside the identity, awaiting signature from others. Once
signing is complete, the transaction can be applied from the staging
area into the identity.

NOTE: staged transactions are not yet synced via the sync system.

Overall, this ia a huge change and it was done with minimal changes to
the interface, which I'm happy about. The system overall is much more
flexible and expansive without adding muh more complexity.